backend/requirements.txt
Django>=5.1,<6.0
PyJWT>=2.8,<3.0
bcrypt>=4.1,<5.0
mysqlclient>=2.2,<3.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.4,<5.0
djangorestframework>=3.15,<4.0
